generate refactor: create subpackage generator
