+++ jwst/associations/__init__.py
@@ -27,7 +27,7 @@ def libpath(filepath):
 from .association import *
 from .association_io import *
 from .exceptions import *
-from .generate import *
+from .generator import *
 from .lib.process_list import *
 from .pool import *
 from .registry import *
