+++ jwst/associations/generator/generate.py
@@ -1,15 +1,15 @@
 import logging
 from timeit import default_timer as timer
 
-from .association import make_timestamp
-from .lib.process_list import (
+from ..association import make_timestamp
+from ..lib.process_list import (
     ListCategory,
     ProcessList,
     ProcessQueueSorted,
     workover_filter
 )
-from .pool import PoolRow
-from ..lib.progress import Bar
+from ..pool import PoolRow
+from ...lib.progress import Bar
 
 # Configure logging
 logger = logging.getLogger(__name__)
