+++ jwst/associations/generator/generate_per_candidate.py
@@ -2,9 +2,9 @@ import collections
 import logging
 
 from .generate import generate
-from .lib.utilities import evaluate
-from .main import CANDIDATE_RULESET, DISCOVER_RULESET, constrain_on_candidates
-from .registry import AssociationRegistry
+from ..lib.utilities import evaluate
+from ..main import CANDIDATE_RULESET, DISCOVER_RULESET, constrain_on_candidates
+from ..registry import AssociationRegistry
 
 # Configure logging
 logger = logging.getLogger(__name__)
